new Place where we store puzzles, Starting working on Puzzles handler, puzzles handler should working with room handler

+++ game/src/World Handling/RoomHandler.h
@@ -40,7 +40,7 @@ class RoomHandler {
 
     int getNrTiles()
     {
-        return tiles.size();
+        return (int)tiles.size();
     }
 
     Tile getTile(int index)
